Moved the api calls into their own class.
